requirements.txt
# Core AI
anthropic>=0.40.0

# Environment
python-dotenv>=1.0.0

# Web UI
streamlit>=1.40.0

# Document Processing
python-docx>=1.1.0
openpyxl>=3.1.2
python-pptx>=1.0.2
PyPDF2>=3.0.1
pdfplumber>=0.11.0
reportlab>=4.2.0
fpdf2>=2.8.1

# Image Processing
Pillow>=11.0.0
img2pdf>=0.5.1

# PDF Conversion
pdf2docx>=0.5.8

# Salesforce
simple-salesforce>=1.12.6

# HTTP & Async
httpx>=0.27.0
aiofiles>=24.1.0
requests>=2.32.0

# File Type Detection
filetype>=1.2.0

# Utilities
loguru>=0.7.2
tqdm>=4.66.0
